feat: value renamed to checked on selection controls
On "Checkbox", "Radio" and "Switch" components, value is renamed to
"checked", and value field is now used to store any value related to the
field. This behavior is similar to HTML and will satisfy use cases like
radio groups.

BREAKING CHANGE: `value` prop in all existing implementations will have to be renamed to "checked".